requirements.txt
Django==1.7.5
argparse==1.2.1
django-cors-headers==1.1.0
django-registration-redux==1.4
djangorestframework==3.3.3
pytz==2016.4
wsgiref==0.1.2
Django==1.7.5
argparse==1.2.1
django-cors-headers==1.1.0
django-registration-redux==1.4
djangorestframework==3.3.3
pytz==2016.4
wsgiref==0.1.2
